Feature/Artwork complete
