+++ components/Button.js
@@ -1,7 +1,9 @@
-export default function Button({ children, buttonstyle, onclick }) {
+export default function Button({
+  children, buttonstyle, onclick, ...attribProps
+}) {
   return (
     <div className="button__container">
-      <button type="button" className={buttonstyle} onClick={onclick}>
+      <button type="button" className={buttonstyle} onClick={onclick} {...attribProps}>
         {children}
       </button>
     </div>
