+++ components/Pagination.js
@@ -1,55 +1,24 @@
 import React, { useState, useEffect } from 'react';
+import useViewport from '../hooks/useViewport';
+import { range } from '../utils/function';
+import Button from './Button';
 
 export default function Pagination({
-  total, perPage, onPrevPage, onNextPage, onPage, className,
+  totalPages,
+  currentPage,
+  onPrevPage,
+  onNextPage,
+  onPage,
+  className,
 }) {
   const btnWidth = 48;
   const countToShow = 3;
-  const totalPages = Math.ceil(total / perPage);
   const paginationClass = className ? `pagination ${className}` : 'pagination';
 
-  const [currentPage, setCurrentPage] = useState(1);
+  const { viewport } = useViewport();
+  const [btnClass, setBtnClass] = useState('');
   const [pages, setPages] = useState([]);
   const [margin, setMargin] = useState(0);
-
-  const prevPage = () => {
-    if (currentPage > 1) {
-      setCurrentPage(currentPage - 1);
-      if (onPrevPage) {
-        onPrevPage();
-      }
-    }
-  };
-
-  const nextPage = () => {
-    if (currentPage < totalPages) {
-      setCurrentPage(currentPage + 1);
-      if (onNextPage) {
-        onNextPage();
-      }
-    }
-  };
-
-  const goToPage = (page) => {
-    setCurrentPage(Math.max(1, Math.min(page, totalPages)));
-    if (onPage) {
-      onPage(page);
-    }
-  };
-
-  const range = (from, to) => {
-    let i = from;
-    const listRange = [];
-
-    while (i <= to) {
-      listRange.push({
-        value: i,
-        index: `page_${i}`,
-      });
-      i += 1;
-    }
-    return listRange;
-  };
 
   const getMargin = () => {
     let newMargin = 0;
@@ -68,39 +37,62 @@ export default function Pagination({
 
   useEffect(() => {
     setPages(range(1, totalPages));
-    setCurrentPage(1);
-  }, [total, perPage]);
+  }, [totalPages]);
 
   useEffect(() => {
     getNumbers();
   }, [currentPage]);
 
+  useEffect(() => {
+    const baseClass = 'button pagination__btn';
+    const newBtnClass = viewport === 'mobile' ? `${baseClass} button--secondary` : `${baseClass} button--bordered`;
+    setBtnClass(newBtnClass);
+  }, [viewport]);
+
   return (
     <div className={paginationClass}>
       <div className="pagination__prev">
         { !(currentPage === 1)
-          && <button className="pagination__btn" type="button" onClick={prevPage}> &lt; </button>}
+          && (
+            <Button buttonstyle={btnClass} onclick={onPrevPage} aria-label="Atrás">
+              <span className="pagination--desktop">&lt;</span>
+              <span className="pagination--mobile">Atrás</span>
+            </Button>
+          )}
       </div>
-      <ul className="pagination__list">
-        { pages.length > 1 && pages.map((page, index) => (
-          <li
-            key={page.index}
-            className="pagination__item"
-            style={index === 0 ? { marginLeft: margin } : {}}
-          >
-            <button
-              type="button"
-              className={`pagination__btn${currentPage === page.value ? ' pagination__btn-active' : ''}`}
-              onClick={() => goToPage(page.value)}
-            >
-              { page.value }
-            </button>
-          </li>
-        )) }
-      </ul>
+      { pages.length > 1 && (
+        <>
+          <ul className="pagination__list pagination__list--desktop">
+            { pages.map((page, index) => (
+              <li
+                key={page.index}
+                className="pagination__item"
+                style={index === 0 ? { marginLeft: margin } : {}}
+              >
+                <Button
+                  buttonstyle={`pagination__btn${currentPage === page.value ? ' pagination__btn-active' : ''}`}
+                  onclick={() => onPage(page.value)}
+                >
+                  { page.value }
+                </Button>
+              </li>
+            )) }
+          </ul>
+          <p className="pagination__list pagination__list--mobile">
+            {currentPage}
+            /
+            {totalPages}
+          </p>
+        </>
+      )}
       <div className="pagination__next">
         { !(currentPage === totalPages)
-          && <button className="pagination__btn" type="button" onClick={nextPage}> &gt; </button>}
+          && (
+            <Button buttonstyle={btnClass} onclick={onNextPage} aria-label="Siguiente">
+              <span className="pagination--desktop">&gt;</span>
+              <span className="pagination--mobile">SIguiente</span>
+            </Button>
+          )}
       </div>
     </div>
   );
