+++ styles/abstracts/_variables.scss
@@ -5,9 +5,14 @@
   --antique-brass: #ca7e6b;
   --pearl-bush: #f0e9e5;
   --paco: #40180f;
+  --gray: #c4c4c4;
+  --light-gray: #e0e0e0;
   --cod-gray: #1b1b1b;
+  --line-gray: #d2d2d2;
+  --semi-gray: #b0b0b0;
   --white: #fff;
   --black: #000;
+  --light-black: #9e9e9e;
   --pink: #ffd4c9;
 
   //Elevations
