+++ styles/components/_artwork.scss
@@ -1,5 +1,5 @@
 .artwork {
-  background-color: #f0e9e5;
+  background-color: var(--pearl-bush);
 
   &__header {
     padding: 35px $mobile-margin;
