+++ styles/components/_cardArtwork.scss
@@ -20,8 +20,8 @@
   }
 
   &__date {
-    border-top: 1px solid #d2d2d2;
-    color: #9e9e9e;
+    border-top: 1px solid var(--line-gray);
+    color: var(--light-black);
   }
 
   &__more {
@@ -54,7 +54,7 @@
     font-size: 1rem;
     line-height: 24px;
     letter-spacing: 0.005em;
-    color: #1b1b1b;
+    color: var(--cod-gray);
   }
 }
