+++ styles/globals.scss
@@ -11,6 +11,7 @@
 @import "./layout/footer";
 
 // components
+@import "./components/button";
 @import "./components/navigation";
 @import "./components/stamp";
 @import "./components/socialMedia";
@@ -22,5 +23,4 @@
 @import "./components/cardArtwork";
 @import "./components/pagination";
 @import "./components/profileCard";
-@import "./components/button";
 @import "./components/profiles";
